+++ client/components/UserColumn/types.tsx
@@ -1,5 +1,5 @@
 import { User } from 'types'
 
 export interface IUserColumnProps {
-  user: Pick<User, 'displayName' | 'mail'>
+  user: Pick<User, 'displayName' | 'mail' | 'photo'>
 }