+++ server/graphql/resolvers/user/types.ts
@@ -6,6 +6,25 @@ import { Subscription } from '../subscription/types'
 import { Role } from '../types'
 
 /**
+ * A type that describes a User photo
+ *
+ * @category GraphQL ObjectType
+ */
+@ObjectType({
+  description: 'A type that describes a User photo',
+  simpleResolvers: true
+})
+export class UserPhoto {
+  @Field({
+    description: 'Base64 representation of the User photo',
+    nullable: true
+  })
+  base64?: string
+}
+
+/**
+ * A type that describes a User
+ *
  * @category GraphQL ObjectType
  */
 @ObjectType({
@@ -56,6 +75,9 @@ export class User {
 
   @Field({ nullable: true })
   provider?: string
+
+  @Field({ nullable: true })
+  photo?: UserPhoto
 
   public create?(user: User): User {
     Object.assign(this, user)